sampling_tests

+++ src/walkjump/cmdline/_sample.py
@@ -17,7 +17,6 @@ def sample(cfg: DictConfig) -> None:
         print(OmegaConf.to_yaml(log_cfg))
 
     hydra.utils.instantiate(cfg.setup)
-    model = hydra.utils.instantiate(cfg.model)
 
     if cfg.device is None:
         device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
@@ -25,21 +24,22 @@ def sample(cfg: DictConfig) -> None:
         print(f"using device {cfg.device}")
         device = torch.device(cfg.device)
 
-    model.to(device)
-
     mask_idxs = instantiate_redesign_mask(cfg.designs.redesign_regions or [])
     seeds = instantiate_seeds(cfg.designs)
-    sample_df = walkjump(
-        seeds,
-        model,
-        delta=cfg.langevin.delta,
-        lipschitz=cfg.langevin.lipschitz,
-        friction=cfg.langevin.friction,
-        steps=cfg.langevin.steps,
-        num_samples=cfg.designs.num_samples,
-        mask_idxs=mask_idxs,
-        chunksize=cfg.designs.chunksize,
-    )
-    sample_df.drop_duplicates(subset=["fv_heavy_aho", "fv_light_aho"], inplace=True)
-    print(f"Writing {len(sample_df)} samples to {cfg.designs.output_csv}")
-    sample_df.to_csv(cfg.designs.output_csv, index=False)
+
+    if not cfg.dryrun:
+        model = hydra.utils.instantiate(cfg.model).to(device)
+        sample_df = walkjump(
+            seeds,
+            model,
+            delta=cfg.langevin.delta,
+            lipschitz=cfg.langevin.lipschitz,
+            friction=cfg.langevin.friction,
+            steps=cfg.langevin.steps,
+            num_samples=cfg.designs.num_samples,
+            mask_idxs=mask_idxs,
+            chunksize=cfg.designs.chunksize,
+        )
+        sample_df.drop_duplicates(subset=["fv_heavy_aho", "fv_light_aho"], inplace=True)
+        print(f"Writing {len(sample_df)} samples to {cfg.designs.output_csv}")
+        sample_df.to_csv(cfg.designs.output_csv, index=False)
